+++ src/main/java/racingcar/domain/GameResultGenerator.java
@@ -1,5 +1,8 @@
 package racingcar.domain;
 
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Cars;
+
 import java.util.List;
 
 public class GameResultGenerator {
