+++ src/main/java/racingcar/domain/RacingCarGameRunner.java
@@ -1,6 +1,8 @@
 package racingcar.domain;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Cars;
 
 import java.util.Arrays;
 import java.util.List;
